animals mejorado

+++ animals/animal.py
@@ -1,5 +1,6 @@
 import numpy as np
 import cv2
+import os
 
 class Animal:
 
@@ -19,7 +20,7 @@ class Cat(Animal):
 		super().__init__(name, color, age, image)
 
 	def sound(self):
-		print('miau')
+		print('MIAU!')
 		
 
 class Dog(Animal):
@@ -28,7 +29,7 @@ class Dog(Animal):
 		super().__init__(name, color, age, image)
 
 	def sound(self):
-		print('guau')
+		print('GUAU!')
 		
 
 class Spider(Animal):
@@ -43,22 +44,22 @@ def main():
 
 	animals = []
 
-	cat_1 = Cat('Nitrito', 'white', '7', '/Users/owner/Documents/Documents/ERICK/project-alfa/animals/images/nitrito.jpeg')
+	cat_1 = Cat('Nitrito', 'white', '7', os.path.dirname(os.path.abspath(__file__))+'/images/nitrito.jpeg')
 	animals.append(cat_1)
-	cat_2 = Cat('Sifu', 'brown', '15', '/Users/owner/Documents/Documents/ERICK/project-alfa/animals/images/sifu.jpg')
+	cat_2 = Cat('Sifu', 'brown', '15', os.path.dirname(os.path.abspath(__file__))+'/images/sifu.jpg')
 	animals.append(cat_2)
-	dog_1 = Dog('Gohan', 'black', '5', '/Users/owner/Documents/Documents/ERICK/project-alfa/animals/images/gohan.jpeg')
+	dog_1 = Dog('Gohan', 'black', '5', os.path.dirname(os.path.abspath(__file__))+'/images/gohan.jpeg')
 	animals.append(dog_1)
-	spider_1 = Spider('Tarantina', 'black', 'unknown', '/Users/owner/Documents/Documents/ERICK/project-alfa/animals/images/tarantina.jpeg')
+	spider_1 = Spider('Tarantina', 'black', 'unknown', os.path.dirname(os.path.abspath(__file__))+'/images/tarantina.jpeg')
 	animals.append(spider_1)
 
 
-	print('******** ANIMALS INFORMATION ********\n')
+	print('****** ANIMALS INFORMATION ******\n')
 	print('LIST:')
 	for x in animals:
 		print(x.name)
 
-	selec = str(input(print('Please choose an animal\n')))
+	selec = str(input('Please choose an animal\n'))
 
 	for x in animals:
 		if selec == x.name:
@@ -66,15 +67,17 @@ def main():
 			print('Name:', x.name)
 			print('Color:', x.color)
 			print('Age:', x.age)
-			print('Sound:')
-			x.sound()
 			print()
+			sound = str(input('Do you want to see {}s sound? (y/n)'.format(x.name)))
+			if sound == 'y':
+				x.sound()
 
-			img = cv2.imread(x.image,1)
-			cv2.namedWindow(x.name,cv2.WINDOW_NORMAL)
-			cv2.imshow(x.name,img)
-			cv2.waitKey(0)
-			cv2.destroyAllWindows()
+			image = str(input('Do you want to see {}s image? (y/n)'.format(x.name)))
+			if image == 'y':
+				img = cv2.imread(x.image,1)
+				cv2.namedWindow(x.name,cv2.WINDOW_NORMAL)
+				cv2.imshow(x.name,img)
+				cv2.waitKey(0)
+				cv2.destroyAllWindows()
 	
-
 main()
